os/bluestore: fix Allocator::allocate() int truncation
An allocation of size 0x80000000 gets truncated to 0.  Larger values can
break things in similar ways.

Introduced in b65bce3a29e529ea24a64f94259e01a9cc466d74.

Fixes: http://tracker.ceph.com/issues/18595

+++ src/os/bluestore/Allocator.h
@@ -39,8 +39,8 @@ public:
 			   uint64_t max_alloc_size, int64_t hint,
 			   AllocExtentVector *extents) = 0;
 
-  int allocate(uint64_t want_size, uint64_t alloc_unit,
-               int64_t hint, AllocExtentVector *extents) {
+  int64_t allocate(uint64_t want_size, uint64_t alloc_unit,
+		   int64_t hint, AllocExtentVector *extents) {
     return allocate(want_size, alloc_unit, want_size, hint, extents);
   }
